Add files via upload
Kiva App Sample Wireframe Pages - A Project for creating a Start Page and Main Page for a Kiva Borrower.  See Project 2 Guidelines and Rubric for the full details of the assignment

Food Scanner App Wireframe - A Project for a creating 4 pages of a wireframe design for a Food App Scanner(Though I did 15 pages).  See Project 1 Guidelines and Rubric for the full details of the assignment